add metod en admin

+++ src/main/java/com/example/registrationlogindemo/repository/ArticleRepository.java
@@ -13,7 +13,9 @@ public interface ArticleRepository extends JpaRepository<Article, Long> {
     List<Article> findAll();
     Optional<Article> findById(Long id);
 
-    void deleteById(Long Id);
+    void deleteById(int Id);
 
     List<Article> findByUser(User usuario);
+
+
 }
